fix: rename json directory to prevent conflicts
