+++ src/otg/assets/__init__.py
