+++ src/otg/assets/data/gwascat_2_gnomad_superpopulation_map.json
