+++ src/otg/assets/schemas/__init__.py
