+++ src/otg/assets/schemas/colocalisation.json
