+++ src/otg/assets/schemas/intervals.json
