+++ src/otg/assets/schemas/studies.json
