+++ src/otg/assets/schemas/study_locus.json
