+++ src/otg/assets/schemas/study_locus_overlap.json
