+++ src/otg/assets/schemas/targets.json
