+++ src/otg/assets/schemas/v2g.json
