+++ src/otg/assets/schemas/variant_annotation.json
