+++ src/otg/assets/schemas/variant_index.json
